--- storage/coop/rchain/key-value-store/build.sbt
@@ -1,7 +1,0 @@
-lazy val root = (project in file("."))
-  .settings(
-    name := "KeyValueStore",
-    scalaVersion := "2.12.2"
-  )
-
-connectInput in run := true
