--- storage/coop/rchain/key-value-store/stores/storeFlat.txt
@@ -1,22 +1,0 @@
-
-A(1) one
-A( 2  ) two
-A(3, 4 ,5) three
-A(3,4,5) four
-A(3,4,5) five
-B(100,101,  100) bee
-B(10,11) yeah
-A(3,5,7) foo
-A(2,4,8) bar
-C(x) keyvariable
-C(1) seeone
-
-// query: A(1), A(2), A(x), A(y)
-A(x) what
-
-// query: B(1,1), B(1,2), B(1,x), B(1,y), B(a,b), B(x,2), B(x,1)
-B(1,2) who
-B(1,x) where
-B(x,2) how
-B(x,y) this
-
