--- storage/coop/rchain/key-value-store/stores/storeNested.txt
@@ -1,16 +1,0 @@
-A(B(1)) one
-A(B(y)) two
-A(1,B(2)) three
-A(y,B(1)) four
-A(1,B(y)) five
-A(y,B(z)) six
-A(B(1),2) seven
-A(B(y),1) eight
-A(B(1),y) nine
-A(B(y),z) ten
-
-A(1,B(2,3),4) eleven
-A(1) twelve
-// matches A(B(1)) or A(B(y))?
-A(x) thirteen
-
