--- storage/coop/rchain/key-value-store/stores/storeRecursive.txt
@@ -1,3 +1,0 @@
-A(B(C(y))) X
-A(B(C(a), D(E(b))), F(c)) Y
-
